NEW: Add optional phpstan execution to PHPSTAN_TEST
PHPStan is a static analyser that may be of use to some repositories.
Notably, I've got a small, passing PHPStan level 1 PR to submit to
framework.

This will include a phpstan call in the PHPSTAN_TEST build, if
phpstan.neon.dist exists.

I've opted not to split it out as a separate build var to reduce the
number of separate builds. This is essentially another form of
sophisticated linting, hence putting it with phpcs. However, it is
linked to a separate variable.